Crate poli_gon

Structs

BfsIterator

Breadth-first traversal iterator for the scene graph.

BoxGeometry

Rectangular cuboid shape.

BufferGeometry

Shape of triangular polygon 3D mesh with compiled buffer data.

Camera

The view point from which 3D scenes are rendered.

DfsIterator

Depth-first traversal iterator for the scene graph.

ForwardRenderer

Forward renderer.

ForwardRendererOptions

Defines the ForwardRenderer’s behavior.

Gpu

Container for several GPU objects used by renderers.

GpuOptions

Parameters for when creating a new GPU adapter.

Mesh

Triangular polygon mesh 3D object.

ObjParseOptions

Configures the ObjParser’s behavior.

ObjParseResult

Successful OBJ file parse result.

ObjParser

Parser for ASCII OBJ files.

Object3d

3D object, which is anything that can have a transformation in 3D space e.g. position, rotation, scale.

Enums

CameraKind

Contains different kinds of cameras.

ObjParseError

Contains OBJ parse errors.

Object3dKind

Contains different kinds of 3D objects.

PowerPreference

Power Preference when choosing a physical adapter.

Traits

Geometry

Generator of common geometry buffers such as vertices and indices. Should be implemented on all geometry parameter structs.