Struct poli_gon::Object3d

pub struct Object3d {
    pub name: RefCell<Option<String>>,
    pub parent: RefCell<Weak<Object3d>>,
    pub children: RefCell<Vec<Rc<Object3d>>>,
    pub local_matrix: RefCell<Matrix4>,
    pub world_matrix: RefCell<Matrix4>,
    pub kind: Object3dKind,
    pub visible: RefCell<bool>,
}

3D object, which is anything that can have a transformation in 3D space e.g. position, rotation, scale.

Fields
name: RefCell<Option<String>>

The name of this 3D object.

parent: RefCell<Weak<Object3d>>

The parent of this 3D object in the 3D scene hierarchy.

children: RefCell<Vec<Rc<Object3d>>>

The children of this 3D object in the 3D scene hierarchy.

local_matrix: RefCell<Matrix4>

The local matrix, which encodes the position, rotation, and scale of this 3D object in model space.

world_matrix: RefCell<Matrix4>

The world matrix, which encodes the position, rotation, and scale of this 3D object in world space.

kind: Object3dKind

The kind of 3D object e.g. mesh, group, camera, light.

visible: RefCell<bool>

Whether or not this 3D object is visible.

Implementations

impl Object3d

pub fn new(kind: Object3dKind) -> Self

Creates a new 3D object of the specified kind at the world origin.

pub fn bfs(object: &Rc<Self>) -> BfsIterator

Returns the breadth-first traversal iterator for the given 3D object, which traverses the scene graph level-by-level starting with the given 3D object itself.

pub fn dfs(object: &Rc<Self>) -> DfsIterator

Returns the depth-first traversal iterator for the given 3D object, which explores along each branch as far as possible before backtracking.

pub fn add(parent: &Rc<Self>, child: &Rc<Self>)

Adds the child object to the parent object and updates the child object’s and its descendants’ world matrices.

pub fn translate(object: &Rc<Self>, v: &Vector3)

Translates the given 3D object in 3D space and updates its descendants’ world matrices.

pub fn rotate(object: &Rc<Self>, q: &Quaternion)

Rotates the given 3D object and updates its descendants’ world matrices.

If you have Euler angles, you can use Quaternion::from.

pub fn scale(object: &Rc<Self>, v: &Vector3)

Scales the given 3D object and updates it descendants’ world matrices.

pub fn update_world_matrix(object: &Rc<Self>, update_descendants: bool)

Updates the world matrix of the given object and (if specified) the world matrices of its descendants.